Refactor styles to styles folder
