+++ client/components/Reviews/Reviews.test.js
@@ -21,7 +21,7 @@ it('should render a reviews list', () => {
 it('should display initial UI', () => {
   render(<Reviews reviews={exampleReviews} />);
 
-  expect(screen.getByRole('option', { name: /relevant/i })).toBeInTheDocument();
+  expect(screen.getByRole('option', { name: /relevance/i })).toBeInTheDocument();
   expect(screen.getByRole('option', { name: /helpful/i })).toBeInTheDocument();
   expect(screen.getByRole('option', { name: /newest/i })).toBeInTheDocument();
