+++ client/components/Reviews/ReviewsList.jsx
@@ -1,22 +1,13 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import styled from 'styled-components';
+import { List } from './styles/ReviewsList.styled';
 import ReviewsListItem from './ReviewsListItem';
 
-const ReviewsList = ({ reviews }) => {
-  const List = styled.div`
-    > :not(:last-child) {
-      border-bottom: 1px solid black;
-    };
-    margin: 25px 0;
-  `;
-
-  return (
-    <List data-testid="reviews-list">
-      {reviews.map((review) => <ReviewsListItem key={review.review_id} review={review} />)}
-    </List>
-  );
-};
+const ReviewsList = ({ reviews }) => (
+  <List data-testid="reviews-list">
+    {reviews.map((review) => <ReviewsListItem key={review.review_id} review={review} />)}
+  </List>
+);
 
 export default ReviewsList;
