+++ client/components/Reviews/styles/ReviewsComponent.styled.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+  display: grid;
+  grid-template-columns: 30% 70%;
+  margin: 0 50px;
+`;
