+++ client/components/Reviews/styles/ReviewsList.styled.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const List = styled.div`
+  > :not(:last-child) {
+    border-bottom: 1px solid black;
+  };
+  margin: 25px 0;
+`;
